changing data from int64_t to float

+++ reworking/src/helpers/predicate.cpp
@@ -6,11 +6,11 @@
 class Predicate
 {
 public:
-    int64_t low;
-    int64_t high;
-    int64_t column;
+    float low;
+    float high;
+    float column;
 
-    Predicate(int64_t low, int64_t high, int64_t column)
+    Predicate(float low, float high, float column)
         : low(low), high(high), column(column){}
     ~Predicate(){}
 };
